Move test to better spot
